requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.5.0
# pydantic-settings is required for BaseSettings in Pydantic v2
pydantic-settings>=2.5.2
python-dotenv>=1.0.1
google-generativeai>=0.7.0
httpx>=0.27.0
tenacity>=8.2.3
